requirements.txt
streamlit
streamlit-authenticator
google-genai
python-dotenv
boto3
requests
beautifulsoup4
Pillow
bcrypt
pandas
